Fixes #14362 - allow passenger to write to /var/www/html/*
Previously, katello could not write to /var/www/html/*. However, during testing
we found that exporting directly to /var/www/html/pub/export is a popular use
case.

This rule allows passenger to write to /var/www/html/*. I successfully compiled
it on el7 and I visually confirmed that the rule exists on el6 as well.